refactor(resolvers): split resolvers in separate files + add some docs
